Sb 1345 rename boosters (#106)
* Replace booster by example

Related to: SB-1345
