+++ .openshiftio/application.yaml
@@ -1,7 +1,7 @@
 apiVersion: v1
 kind: Template
 metadata:
-  name: spring-boot-rest-http
+  name: rest-http
   annotations:
     iconClass: icon-spring
     tags: spring-boot, rest, java, microservice
@@ -49,7 +49,7 @@ objects:
 - apiVersion: v1
   kind: ImageStream
   metadata:
-    name: spring-boot-rest-http
+    name: rest-http
   spec: {}
 - apiVersion: v1
   kind: ImageStream
@@ -64,12 +64,12 @@ objects:
 - apiVersion: v1
   kind: BuildConfig
   metadata:
-    name: spring-boot-rest-http
+    name: rest-http
   spec:
     output:
       to:
         kind: ImageStreamTag
-        name: spring-boot-rest-http:BOOSTER_VERSION
+        name: rest-http:BOOSTER_VERSION
     postCommit: {}
     resources: {}
     source:
@@ -106,11 +106,11 @@ objects:
   kind: Service
   metadata:
     labels:
-      app: spring-boot-rest-http
+      app: rest-http
       provider: snowdrop
       version: "BOOSTER_VERSION"
-      group: io.openshift.booster
-    name: spring-boot-rest-http
+      group: dev.snowdrop.example
+    name: rest-http
   spec:
     ports:
     - name: http
@@ -118,24 +118,24 @@ objects:
       protocol: TCP
       targetPort: 8080
     selector:
-      app: spring-boot-rest-http
+      app: rest-http
       provider: snowdrop
-      group: io.openshift.booster
+      group: dev.snowdrop.example
 - apiVersion: v1
   kind: DeploymentConfig
   metadata:
     labels:
-      app: spring-boot-rest-http
+      app: rest-http
       provider: snowdrop
       version: "BOOSTER_VERSION"
-      group: io.openshift.booster
-    name: spring-boot-rest-http
+      group: dev.snowdrop.example
+    name: rest-http
   spec:
     replicas: 1
     selector:
-      app: spring-boot-rest-http
+      app: rest-http
       provider: snowdrop
-      group: io.openshift.booster
+      group: dev.snowdrop.example
     strategy:
       rollingParams:
         timeoutSeconds: 3600
@@ -143,10 +143,10 @@ objects:
     template:
       metadata:
         labels:
-          app: spring-boot-rest-http
+          app: rest-http
           provider: snowdrop
           version: "BOOSTER_VERSION"
-          group: io.openshift.booster
+          group: dev.snowdrop.example
       spec:
         containers:
         - env:
@@ -154,7 +154,7 @@ objects:
             valueFrom:
               fieldRef:
                 fieldPath: metadata.namespace
-          image: spring-boot-rest-http:BOOSTER_VERSION
+          image: rest-http:BOOSTER_VERSION
           imagePullPolicy: IfNotPresent
           name: spring-boot
           ports:
@@ -186,20 +186,20 @@ objects:
         - spring-boot
         from:
           kind: ImageStreamTag
-          name: spring-boot-rest-http:BOOSTER_VERSION
+          name: rest-http:BOOSTER_VERSION
       type: ImageChange
 - apiVersion: v1
   kind: Route
   metadata:
     labels:
-      app: spring-boot-rest-http
+      app: rest-http
       provider: snowdrop
       version: "BOOSTER_VERSION"
-      group: io.openshift.booster
-    name: spring-boot-rest-http
+      group: dev.snowdrop.example
+    name: rest-http
   spec:
     port:
       targetPort: 8080
     to:
       kind: Service
-      name: spring-boot-rest-http
+      name: rest-http
