+++ README.md
@@ -1,3 +1,3 @@
-[![CircleCI](https://circleci.com/gh/snowdrop/spring-boot-http-booster/tree/master.svg?style=shield)](https://circleci.com/gh/snowdrop/spring-boot-http-booster/tree/master)
+[![CircleCI](https://circleci.com/gh/snowdrop/rest-http-example/tree/master.svg?style=shield)](https://circleci.com/gh/snowdrop/rest-http-example/tree/master)
 
 https://appdev.openshift.io/docs/spring-boot-runtime.html#mission-http-api-spring-boot
