+++ pom.xml
@@ -16,11 +16,11 @@
 -->
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>io.openshift.booster</groupId>
-  <artifactId>spring-boot-rest-http</artifactId>
+  <groupId>dev.snowdrop.example</groupId>
+  <artifactId>rest-http</artifactId>
   <version>2.1.3-3-SNAPSHOT</version>
-  <name>Spring Boot - HTTP Booster</name>
-  <description>Spring Boot - HTTP Booster</description>
+  <name>Spring Boot - HTTP Example</name>
+  <description>Spring Boot - HTTP Example</description>
   <properties>
     <maven.compiler.source>1.8</maven.compiler.source>
     <maven.compiler.target>1.8</maven.compiler.target>
