+++ src/main/java/dev/snowdrop/example/ExampleApplication.java
@@ -1,36 +1,34 @@
 /*
  * Copyright 2016-2017 Red Hat, Inc, and individual contributors.
- *
+ * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * <p>
  * http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package dev.snowdrop.example;
 
-package io.openshift.booster;
+import com.fasterxml.jackson.jaxrs.json.JacksonJsonProvider;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
 
-import java.net.URL;
-import org.arquillian.cube.openshift.impl.enricher.AwaitRoute;
-import org.arquillian.cube.openshift.impl.enricher.RouteURL;
-import org.jboss.arquillian.junit.Arquillian;
-import org.junit.runner.RunWith;
+@SpringBootApplication
+public class ExampleApplication {
 
-@RunWith(Arquillian.class)
-public class OpenShiftIT extends AbstractBoosterApplicationTest {
+    public static void main(String[] args) {
+        SpringApplication.run(ExampleApplication.class, args);
+    }
 
-    @AwaitRoute(path = "/actuator/health")
-    @RouteURL("${app.name}")
-    private URL baseURL;
-
-    @Override
-    public String baseURI() {
-        return baseURL.toString();
+    @Bean
+    public JacksonJsonProvider jsonProvider() {
+        return new JacksonJsonProvider();
     }
 }
