+++ src/main/java/dev/snowdrop/example/service/Greeting.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.openshift.booster.service;
+package dev.snowdrop.example.service;
 
 // tag::snippet-greeting[]
 public class Greeting {
