+++ src/main/resources/static/index.html
@@ -10,7 +10,7 @@
 <div class="container">
     <div class="row">
         <div class="sect1">
-<h2 id="_http_booster">HTTP Booster</h2>
+<h2 id="_http_booster">HTTP Example</h2>
 <div class="sectionbody">
 <div class="paragraph">
 <p>An example of simple mapping of a business operation to a remote endpoint. By taking this approach, clients use the HTTP protocol as a transport mechanism to call services. Application engineers define the APIs using a broad interpretation of the REST fundamentals, encouraging freedom in design and quick prototyping.</p>
