+++ src/test/java/dev/snowdrop/example/OpenShiftIT.java
@@ -1,34 +1,36 @@
 /*
  * Copyright 2016-2017 Red Hat, Inc, and individual contributors.
- * <p>
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * <p>
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.openshift.booster;
 
-import com.fasterxml.jackson.jaxrs.json.JacksonJsonProvider;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
+package dev.snowdrop.example;
 
-@SpringBootApplication
-public class BoosterApplication {
+import java.net.URL;
+import org.arquillian.cube.openshift.impl.enricher.AwaitRoute;
+import org.arquillian.cube.openshift.impl.enricher.RouteURL;
+import org.jboss.arquillian.junit.Arquillian;
+import org.junit.runner.RunWith;
 
-    public static void main(String[] args) {
-        SpringApplication.run(BoosterApplication.class, args);
-    }
+@RunWith(Arquillian.class)
+public class OpenShiftIT extends AbstractExampleApplicationTest {
 
-    @Bean
-    public JacksonJsonProvider jsonProvider() {
-        return new JacksonJsonProvider();
+    @AwaitRoute(path = "/actuator/health")
+    @RouteURL("${app.name}")
+    private URL baseURL;
+
+    @Override
+    public String baseURI() {
+        return baseURL.toString();
     }
 }
